# analysis-service/src/test/resources/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Chromamon Analysis Service API
  description: API Documentation for Analysis Service for the Chromamon project
  contact:
    name: Paulo Possatto
    url: https://github.com/Paulo-Possatto
  license:
    name: MIT License
    url: https://github.com/Paulo-Possatto/Chromamon/blob/main/LICENSE
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local environment for testing
tags:
- name: Get Analysis
  description: All resources related to retrieve analysis data from the database
- name: Post Analysis
  description: All resources related to save analysis data into the database
- name: Put Analysis
  description: All resources related to modify analysis data saved into the database
- name: Delete Analysis
  description: All resources related to erase analysis data into the database
paths:
  /keepalive:
    get:
      tags:
      - keep-alive-controller
      operationId: keepAlive
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /api/analyses:
    get:
      tags:
      - Get Analysis
      summary: Get a list of all analyses
      description: |
        Returns a list of all the analyses stored in the database.

        Roles allowed to access this resource:
        - **ADMIN**
        - **USER**
        - **ENGINEER**
        - **CHEMIST**
        - **LAB_ANALYST**
      operationId: getAnalyses
      parameters:
      - name: page
        in: query
        description: The page number
        schema:
          type: integer
          default: 0
      - name: size
        in: query
        description: Number of items per page
        schema:
          type: integer
          default: 20
      - name: sort
        in: query
        description: Sorting criteria
        schema:
          type: string
      - name: X-API-Version
        in: header
        description: The version of the analyses API
        required: true
        schema:
          type: string
      - name: X-Chroma-Token
        in: header
        description: The JWT token used to access this resource
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Give a page of analyses, the amount can be modified in the\
            \ URI params"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analyses-Responses"
              examples:
                Analyses Response:
                  description: The response object that lists all the analyses
                  value:
                    analyses:
                    - identifier: AN_3102554574
                      transformerSerialNumber: TRF-132kV-2023-ABC123
                      analysisDateTime: 25-06-2025 15:19:38
                      labAnalysisDateTime: 26-06-2025 07:15:11
                      chromatography:
                        hydrogen: "0,15"
                        methane: "0,21"
                        ethane: "0,09"
                        acetylene: "0,11"
                        carbonMonoxide: "4,45"
                        carbonDioxide: "915,41"
                        createdDateTime: 26-07-2025 18:35:06
                        lastUpdateDateTime: 26-07-2025 18:35:06
                        userCreated: AA00AA
                        userLastUpdated: AA00AA
                      oilType: Mineral Oil
                      observation:
                        sampleCondition: Presence of particles
                        gasExtractionMethod: Direct
                        modelUsed: GC Agilent 7890B
                        createdDateTime: 26-07-2025 18:35:06
                        lastUpdateDateTime: 26-07-2025 18:35:06
                        userCreated: AA00AA
                        userLastUpdated: AA00AA
                      furfuralAnalysis: "0,50"
                      oilHumidity: "15,00"
                      createdDateTime: 26-07-2025 18:35:06
                      lastUpdateDateTime: 26-07-2025 18:35:06
                      userCreated: AA00AA
                      userLastUpdated: AA00AA
                    - identifier: AN_5069214547
                      transformerSerialNumber: TRF-220kV-2023-XYZ789
                      analysisDateTime: 26-06-2025 06:45:22
                      labAnalysisDateTime: 27-06-2025 08:30:05
                      chromatography:
                        hydrogen: "0,22"
                        methane: "0,35"
                        ethane: "0,12"
                        acetylene: "0,08"
                        carbonMonoxide: "3,75"
                        carbonDioxide: "1.020,10"
                        createdDateTime: 26-07-2025 18:35:06
                        lastUpdateDateTime: 26-07-2025 18:35:06
                        userCreated: AA00AA
                        userLastUpdated: AA00AA
                      oilType: Vegetal Oil
                      observation:
                        sampleCondition: Clear with slight discoloration
                        gasExtractionMethod: Manual Pump
                        modelUsed: GC Shimadzu 2030
                        createdDateTime: 26-07-2025 18:35:06
                        lastUpdateDateTime: 26-07-2025 18:35:06
                        userCreated: AA00AA
                        userLastUpdated: AA00AA
                      furfuralAnalysis: "0,30"
                      oilHumidity: "12,30"
                      createdDateTime: 26-07-2025 18:35:06
                      lastUpdateDateTime: 26-07-2025 18:35:06
                      userCreated: AA00AA
                      userLastUpdated: AA00AA
                    - identifier: AN_8369902135
                      transformerSerialNumber: TRF-66kV-2024-DEF456
                      analysisDateTime: 27-06-2025 12:30:10
                      labAnalysisDateTime: 28-06-2025 06:45:33
                      chromatography:
                        hydrogen: "0,18"
                        methane: "0,28"
                        ethane: "0,11"
                        acetylene: "0,15"
                        carbonMonoxide: "5,20"
                        carbonDioxide: "980,75"
                        createdDateTime: 26-07-2025 18:35:06
                        lastUpdateDateTime: 26-07-2025 18:35:06
                        userCreated: AA00AA
                        userLastUpdated: AA00AA
                      oilType: Silicon Oil
                      observation:
                        sampleCondition: Normal appearance
                        gasExtractionMethod: Electric Pump
                        modelUsed: GC PerkinElmer Clarus 680
                        createdDateTime: 26-07-2025 18:35:06
                        lastUpdateDateTime: 26-07-2025 18:35:06
                        userCreated: AA00AA
                        userLastUpdated: AA00AA
                      furfuralAnalysis: "0,70"
                      oilHumidity: "18,72"
                      createdDateTime: 26-07-2025 18:35:06
                      lastUpdateDateTime: 26-07-2025 18:35:06
                      userCreated: AA00AA
                      userLastUpdated: AA00AA
                    currentPage: 0
                    totalItems: 3
                    totalPages: 1
        "204":
          description: "No analyses are stored in the database, so nothing returned"
        "400":
          description: Some user input that's not been validated by the system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error-Response"
              examples:
                Invalid Token Response:
                  description: The header parameter 'X-Chroma-Token' is invalid
                  value:
                    statusCode: 400
                    message: Bad Request
                    details: Header parameter 'X-Chroma-Token' has a invalid value
                    timestamp: 22-07-2025 10:23:30
                Empty Token Response:
                  description: Header parameter 'X-Chroma-Token' is not present in
                    the request
                  value:
                    statusCode: 400
                    message: Bad Request
                    details: Header parameter 'X-Chroma-Token' must not be empty
                    timestamp: 22-07-2025 10:23:30
                Empty API Version Response:
                  description: The header parameter 'X-API-Version' is not present
                    in the request
                  value:
                    statusCode: 400
                    message: Bad Request
                    details: Header parameter 'X-API-Version' must not be empty
                    timestamp: 22-07-2025 10:23:30
                Invalid API Version Format Response:
                  description: The header parameter 'X-API-Version' is with a wrong
                    format
                  value:
                    statusCode: 400
                    message: Bad Request
                    details: "'X-API-Version' must begin with 'V', followed by the\
                      \ version number"
                    timestamp: 22-07-2025 10:23:30
                Invalid sort param:
                  description: The property used to sort does not exist in the entity
                    object
                  value:
                    statusCode: 400
                    message: Bad Request
                    details: No property 'abc' found for type 'AnalysisEntity'
                    timestamp: 26-07-2025 20:56:52
        "401":
          description: The Header param 'X-Chroma-Token' is not being validated correctly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error-Response"
              examples:
                No Claims Response:
                  description: "'X-Chroma-Token' is provided, but without payload"
                  value:
                    statusCode: 401
                    message: Unauthorized
                    details: Header parameter 'X-Chroma-Token' has no claims
                    timestamp: 22-07-2025 10:23:30
                Expired Token Response:
                  description: "Header parameter 'X-Chroma-Token' is provided, but\
                    \ it's expired"
                  value:
                    statusCode: 401
                    message: Unauthorized
                    details: "Header parameter 'X-Chroma-Token' is expired, please\
                      \ refresh it"
                    timestamp: 22-07-2025 10:23:30
                Wrong Issuer Response:
                  description: The parameter 'X-Chroma-Token' was not issued by the
                    application
                  value:
                    statusCode: 401
                    message: Unauthorized
                    details: Header parameter 'X-Chroma-Token' was not issued by the
                      application
                    timestamp: 22-07-2025 10:23:30
                No Roles Response:
                  description: The header parameter 'X-Chroma-Token' has no roles
                    in its claim set
                  value:
                    statusCode: 401
                    message: Unauthorized
                    details: Header parameter 'X-Chroma-Token' does not contain roles
                      for the user
                    timestamp: 22-07-2025 10:23:30
        "403":
          description: The user has no permission to this request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error-Response"
              examples:
                No Permission Response:
                  description: The user has no role allowed to access the endpoint
                  value:
                    statusCode: 403
                    message: Forbidden
                    details: The user has no permission to access this resource
                    timestamp: 22-07-2025 10:23:30
        "404":
          description: The user sent information that was not found in the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error-Response"
              examples:
                No API Version Found Response:
                  description: The requested API version was not found nor implemented
                  value:
                    statusCode: 404
                    message: Not Found
                    details: API Version not supported
                    timestamp: 22-07-2025 10:23:30
        "500":
          description: Something unhandled happened in the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error-Response"
              examples:
                Internal Server Error Response:
                  description: "Some generic Exception occurred, it's necessary to\
                    \ check the logs to find out where and why this error happened"
                  value:
                    statusCode: 500
                    message: Internal Server Error
                    details: Something happened. Please contact the user admin to
                      check logs
                    timestamp: 22-07-2025 10:23:30
components:
  schemas:
    Analyses-Responses:
      type: object
      properties:
        analyses:
          type: array
          items:
            $ref: "#/components/schemas/Analysis-Response"
        currentPage:
          type: integer
          description: The selected page of contents
          format: int32
          example: 1
        totalItems:
          type: number
          description: The amount of items in the page
          example: 3
        totalPages:
          type: integer
          description: The total number of pages for all the elements
          format: int32
          example: 5
      description: The response object that contains a list of all analyses
    Analysis-Response:
      type: object
      properties:
        identifier:
          type: string
          description: "The identifier of the analysis, starting with 'AN_' followed\
            \ by 10 numbers"
          example: AN_5168350147
        transformerSerialNumber:
          type: string
          description: The serial number of the transformer that was analysed
          example: TRF-132kV-2023-ABC123
        analysisDateTime:
          type: string
          description: The date and time the analysis was carried out
          example: 25-06-2025 17:19:38
        labAnalysisDateTime:
          type: string
          description: The time the analysis was started in the laboratory
          example: 26-06-2025 09:15:11
        chromatography:
          $ref: "#/components/schemas/Chromatography-Response"
        oilType:
          type: string
          description: The type of isolating oil used in the transformer
          example: Mineral Oil
        observation:
          $ref: "#/components/schemas/Observation-Response"
        furfuralAnalysis:
          type: string
          description: The value that determines the condition of the isolating paper
            (value in mg/L)
          example: "0.15"
        oilHumidity:
          type: string
          description: The percentage of humidity detected in the isolating oil
          example: "15.0"
        createdDateTime:
          type: string
          description: The date-time when the data was inserted in the database
          example: 22-07-2025 15:43:25
        lastUpdateDateTime:
          type: string
          description: The date-time when the data was last updated in the database
          example: 22-07-2025 15:43:25
        userCreated:
          type: string
          description: "The subject that created the data, took from the from the\
            \ JWT token"
          example: AA00AA
        userLastUpdated:
          type: string
          description: "The subject that last updated the data, took from the from\
            \ the JWT token"
          example: AA00AA
      description: The object that contains the information about a single analysis
    Chromatography-Response:
      type: object
      properties:
        hydrogen:
          type: string
          description: The concentration (in ppm) of hydrogen gas
          example: "0.18"
        methane:
          type: string
          description: The concentration (in ppm) of methane gas
          example: "0.28"
        ethane:
          type: string
          description: The concentration (in ppm) of ethane gas
          example: "0.11"
        acetylene:
          type: string
          description: The concentration (in ppm) of acetylene gas
          example: "0.15"
        carbonMonoxide:
          type: string
          description: The concentration (in ppm) of carbon monoxide gas
          example: "5.2"
        carbonDioxide:
          type: string
          description: The concentration (in ppm) of carbon dioxide gas
          example: "980.75"
        createdDateTime:
          type: string
          description: The date-time when the data was inserted in the database
          example: 22-07-2025 15:43:25
        lastUpdateDateTime:
          type: string
          description: The date-time when the data was last updated in the database
          example: 22-07-2025 15:43:25
        userCreated:
          type: string
          description: "The subject that created the data, took from the from the\
            \ JWT token"
          example: AA00AA
        userLastUpdated:
          type: string
          description: "The subject that last updated the data, took from the from\
            \ the JWT token"
          example: AA00AA
      description: The object that contains the chromatographic information
    Observation-Response:
      type: object
      properties:
        sampleCondition:
          type: string
          description: The description of the sample condition
          example: Presence of particles
        gasExtractionMethod:
          type: string
          description: The method in which the gas was extracted
          example: Vacuum
        modelUsed:
          type: string
          description: The model of the tool used to extract the oil for the analysis
          example: GC Agilent 7890B
        createdDateTime:
          type: string
          description: The date-time when the data was inserted in the database
          example: 22-07-2025 15:43:25
        lastUpdateDateTime:
          type: string
          description: The date-time when the data was last updated in the database
          example: 22-07-2025 15:43:25
        userCreated:
          type: string
          description: "The subject that created the data, took from the from the\
            \ JWT token"
          example: AA00AA
        userLastUpdated:
          type: string
          description: "The subject that last updated the data, took from the from\
            \ the JWT token"
          example: AA00AA
      description: The object that contains information about the observations while
        analysing
    Error-Response:
      type: object
      properties:
        statusCode:
          type: integer
          description: The HTTP Status Code of the error
          format: int32
          example: 404
        message:
          type: string
          description: "The HTTP Status message, saying what the code means"
          example: Not Found
        details:
          type: string
          description: The details of what caused the error
          example: API Version not supported
        timestamp:
          type: string
          description: The date and time of when the error occurred
          example: 22-07-2025 10:23:30
      description: The object that contains the information about the error